requirements.txt
telethon
pandas
python-dotenv
matplotlib
seaborn
numpy
notebook
jupyter
dbt-postgres
psycopg2-binary # For PostgreSQL connection